renaming the domain folder to model
