+++ 4MD/src/main/java/com/infoshareacademy/four_md/Application.java
@@ -1,12 +1,7 @@
 package com.infoshareacademy.four_md;
 
-import com.infoshareacademy.four_md.domain.*;
 import org.springframework.boot.SpringApplication;
-import org.springframework.boot.SpringBootConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-import java.util.ArrayList;
-import java.util.List;
 
 @SpringBootApplication
 public class Application {
