+++ 4MD/src/main/java/com/infoshareacademy/four_md/model/Ingredients.java
@@ -1,4 +1,4 @@
-package com.infoshareacademy.four_md.domain;
+package com.infoshareacademy.four_md.model;
 
 public class Ingredients {
     private String name;
