+++ 4MD/src/main/java/com/infoshareacademy/four_md/model/Ratings.java
@@ -1,4 +1,4 @@
-package com.infoshareacademy.four_md.domain;
+package com.infoshareacademy.four_md.model;
 
 public enum Ratings {
     R0(0), R1(1), R2(2), R3(3), R4(4), R5(5);
